requirements.txt
Django>=5.0.0
django-allauth>=0.57.0
psycopg2-binary>=2.9.9
pypdf>=6.6.2
pdfplumber>=0.11.0
PyYAML>=6.0.1
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
Pillow>=10.0.0
